+++ src/civis/io/_utils.py
@@ -1,4 +1,11 @@
+from __future__ import annotations
+
+import os
 import uuid
+from typing import TypeAlias
+
+
+TypePathLike: TypeAlias = str | bytes | os.PathLike
 
 
 def maybe_get_random_name(name):